rename test_instance to test_default and test_default to test_all while adding some env variables for images
